Revert "implementation dependency injection for notifications"
