+++ app/Providers/AppServiceProvider.php
@@ -2,10 +2,6 @@
 
 namespace App\Providers;
 
-use App\Services\Notifications\ModeratorNotifier;
-use App\Services\Notifications\NotificationChannel;
-use App\Services\Notifications\TwitterNotifier;
-use App\Services\Notifications\UserNotifier;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\ServiceProvider;
 
@@ -16,13 +12,7 @@ class AppServiceProvider extends ServiceProvider
      */
     public function register(): void
     {
-        $this->app->bind(NotificationChannel::class, function () {
-            return [
-                new TwitterNotifier,
-                new ModeratorNotifier,
-                new UserNotifier,
-            ];
-        });
+        //
     }
 
     /**
